refactor: playlist 조회 기능 AI 코드리뷰 피드백 반영
