+++ mopl-api/src/main/java/io/mopl/api/playlist/dto/PlaylistUpdateRequest.java
@@ -10,7 +10,6 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class PlaylistUpdateRequest {
-
   private String title;
   private String description;
 }
