+++ mopl-api/src/main/java/io/mopl/api/playlist/repository/PlaylistSubscriptionRepository.java
@@ -10,7 +10,6 @@ import org.springframework.stereotype.Repository;
 import io.mopl.api.playlist.domain.PlaylistSubscription;
 import io.mopl.api.playlist.domain.PlaylistSubscriptionId;
 
-@Repository
 public interface PlaylistSubscriptionRepository extends JpaRepository<PlaylistSubscription, PlaylistSubscriptionId> {
 
 	/**
